# example/axum_backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "axum_backend",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/bank-stream": {
      "post": {
        "tags": [
          "bank_stream"
        ],
        "operationId": "dispatch",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BankAccountCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Command dispatched successfully"
          },
          "403": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Domain validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankAccountError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "BankAccountCommand": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "user_id",
              "event"
            ],
            "properties": {
              "event": {
                "type": "string",
                "enum": [
                  "OpenAccount"
                ]
              },
              "user_id": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "amount",
              "account_id",
              "event"
            ],
            "properties": {
              "account_id": {
                "type": "string"
              },
              "amount": {
                "type": "integer",
                "format": "int32"
              },
              "event": {
                "type": "string",
                "enum": [
                  "DepositFunds"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "amount",
              "account_id",
              "event"
            ],
            "properties": {
              "account_id": {
                "type": "string"
              },
              "amount": {
                "type": "integer",
                "format": "int32"
              },
              "event": {
                "type": "string",
                "enum": [
                  "WithdrawFunds"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "user_id",
              "account_id",
              "event"
            ],
            "properties": {
              "account_id": {
                "type": "string"
              },
              "event": {
                "type": "string",
                "enum": [
                  "ChangeOwner"
                ]
              },
              "user_id": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "withdrawal_id",
              "account_id",
              "event"
            ],
            "properties": {
              "account_id": {
                "type": "string"
              },
              "event": {
                "type": "string",
                "enum": [
                  "DisputeWithdrawal"
                ]
              },
              "withdrawal_id": {
                "type": "string"
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "event"
        }
      },
      "BankAccountError": {
        "type": "string",
        "enum": [
          "NotEnoughFunds",
          "AlreadyOpened",
          "AccountNotFound"
        ]
      }
    }
  }
}